refactor(artifacts): extract compatibility owner
