+++ src/main/artifacts/storage-layout.ts
@@ -2,6 +2,14 @@ const ARTIFACTS_DIR = 'artifacts'
 const PENDING_DIR = '.pending'
 const METADATA_DIR = '.metadata'
 const CURRENT_RUN_FILE = 'current-run.json'
+const RUNS_DIR = '.runs'
 const SAFE_SEGMENT_PATTERN = /^[A-Za-z0-9][A-Za-z0-9._-]*$/
 
-export { ARTIFACTS_DIR, CURRENT_RUN_FILE, METADATA_DIR, PENDING_DIR, SAFE_SEGMENT_PATTERN }
+export {
+  ARTIFACTS_DIR,
+  CURRENT_RUN_FILE,
+  METADATA_DIR,
+  PENDING_DIR,
+  RUNS_DIR,
+  SAFE_SEGMENT_PATTERN
+}
